Fixed long standing bug: charts didn't show on initial install
Added line to application startup (for those activities which might
be launched from outside of this app) to commit default preferences
to the preference manager. This has been a VERY longstanding issue
in the app, which caused charts to not show until the settings
screen was visited at least once (even if no settings were changed).